Move cache purging to service
